backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1

# ── Auth ──────────────────────────────────────────────────────
bcrypt==4.1.3

# ── ML / Deep Learning ────────────────────────────────────────
# torch + torchvision installed separately in Dockerfile (CPU build)
timm==0.9.16

# ── GradCAM++ ─────────────────────────────────────────────────
grad-cam==1.5.0

# ── Image Processing ──────────────────────────────────────────
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80

# ── PDF Generation ────────────────────────────────────────────
reportlab==4.2.0

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
gunicorn==22.0.0
